refactor(app): compose project session rows
